Crypto.Cipher.Spritz

Description

The original paper for the Spritz cipher can be found here:

https://people.csail.mit.edu/rivest/pubs/RS14.pdf

This package provides a Haskell implementation of the pseudocode in the paper listed above. It intends to be a direct implementation of the cipher, so we rely heavily on use of the State monad. We also make heavy use of the lens library's combinators, internally, to ease our use of State.

Lastly, I must give a shout-out to spritzjs for the bitmasking parts and for existing so that I had something to test against when I was done.

State/Lenses

data SpritzState

Constructors

SpritzState

Fields

_i :: Int

_j :: Int

_k :: Int

_z :: Int

_w :: Int

_a :: Int

_s :: Vector Int

_n :: Int

i :: Lens' SpritzState Int

j :: Lens' SpritzState Int

k :: Lens' SpritzState Int

z :: Lens' SpritzState Int

w :: Lens' SpritzState Int

a :: Lens' SpritzState Int

s :: Lens' SpritzState (Vector Int)

n :: Lens' SpritzState Int

Spritz basic functions

initializeState :: Int -> SpritzState

absorb :: Vector Int -> State SpritzState ()

absorbByte :: Int -> State SpritzState ()

absorbNibble :: Int -> State SpritzState ()

absorbStop :: State SpritzState ()

shuffle :: State SpritzState ()

whip :: Int -> State SpritzState ()

crush :: State SpritzState ()

squeeze :: Int -> State SpritzState (Vector Int)

drip :: State SpritzState Int

update :: State SpritzState ()

output :: State SpritzState Int

Helper functions

low :: (Bits a, Num a, Show a) => a -> a

high :: (Bits a, Num a, Show a) => a -> a

plusmod :: Integral a => a -> a -> a -> a

submod :: Integral a => a -> a -> a -> a

swap :: Int -> Int -> State SpritzState ()

Swap two elements given indices of S.

Making use of everything

Hashing

hash :: Vector Int -> Int -> SpritzState -> Vector Int

Produces an r-byte hash of the input message.

Encyrption

encrypt

Arguments

:: Vector Int
The key.

-> Vector Int
The decrypted message.

-> SpritzState
Starting state.

-> Vector Int

decrypt

Arguments

:: Vector Int
The key.

-> Vector Int
The encrypted message.

-> SpritzState
Starting state.

-> Vector Int

keySetup

Arguments

:: Int
Our N value. 256 in the paper.

-> Vector Int
The key.

-> State SpritzState ()

Used in the paper at the top of encrypt* and decrypt, but not used by default in this library. Still, we provide it in case it's needed.

keySetup n' k' = put (initializeState n') >> absorb k'